refactor: move multilevelToSankey helper to sankeyUtils.ts
